Kdesktop 1747 New communication layer (step 1)

+++ src/libsyncengine/jobs/network/abstractnetworkjob.h
@@ -76,7 +76,7 @@ class AbstractNetworkJob : public AbstractJob {
         void getStringFromStream(std::istream &inputStream, std::string &res);
 
         std::string _httpMethod;
-        uint64_t _apiVersion{2};
+        uint8_t _apiVersion{2};
         std::string _data;
         Poco::Net::HTTPResponse _resHttp;
         int _customTimeout = 0;
